# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"posts": {
						"items": {
							"$ref": "#/components/schemas/PostModel"
						},
						"type": "array"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/CommentModel"
						},
						"type": "array"
					},
					"createDateColumn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"posts",
					"comments",
					"createDateColumn",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PostModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"userID": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/UserModel"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/CommentModel"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"content",
					"userID",
					"user",
					"comments",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CommentModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/UserModel"
					},
					"postId": {
						"type": "number",
						"format": "double"
					},
					"post": {
						"$ref": "#/components/schemas/PostModel"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"content",
					"userId",
					"user",
					"postId",
					"post",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IPostPayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"userID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"content",
					"userID"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ICommentPayload": {
				"properties": {
					"content": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"postId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"content",
					"userId",
					"postId"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "econfi_v1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserModel"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts": {
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PostModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostModel"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPostPayload"
							}
						}
					}
				}
			}
		},
		"/posts/{id}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PostModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments": {
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CommentModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentModel"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICommentPayload"
							}
						}
					}
				}
			}
		},
		"/comments/{id}": {
			"get": {
				"operationId": "GetComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CommentModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}