# src/routes/UserRouter.ts
import express from "express";
import UserController from "../controllers/UserController";

const router = express.Router();

router.get('/', async (_req, res) => {
  const controller = new UserController();
  const response = await controller.getUsers();

  return res.send(response);
});

router.post('/', async (req, res) => {
  const controller = new UserController();
  const response = await controller.createUser(req.body);

  return res.send(response);
});

router.get('/:id', async (req, res) => {
  
  try {
    const controller = new UserController();
    const response = await controller.getUser(req.params.id);

    if (!response) return res.status(404).send({ message: 'No user found' });

    return res.send(response);

  } catch (error) {
    res.status(400).send({ message: 'error' });
  }

  
});

export default router;